finish package script usage and winpthreads and libgcc fixage.

+++ scripts/package.sh
@@ -3,19 +3,27 @@
 package()
 (
   host="$1"
-  packagename="$2"
+  projectwithversion="$2"
   if [ -z "$3" ]
   then
-    makeinstallargs="install-strip"
+    packagesuffix=
   else
-    makeinstallargs="$3"
+    packagesuffix="$3"
   fi
   
-  logdir="$_CROSS_LOG_DIR/$host/$packagename"
+  packagename="$host-$projectwithversion$packagesuffix"
+
+  logdir="$_CROSS_LOG_DIR/$host/$projectwithversion"
+  mkdir -p "$logdir"
+  
+  if [ -f "$_CROSS_PACKAGE_DIR/$packagename$_CROSS_COMPRESS_EXT" ]
+  then
+    exit 0;
+  fi
 
   printf ">>> Packaging $packagename...\n"
   cd "$_CROSS_STAGE_INSTALL_DIR"
   find . -name \*.la -exec rm -f {} \;
-  $_CROSS_COMPRESS_TAR "$_CROSS_PACKAGE_DIR/$packagename$_CROSS_COMPRESS_EXT" ./* > "$logdir/package$buildstep.log" 2>&1 \
-    || { printf "Failure packaging $project$buildstep. Check $logdir/install$buildstep.log for details.\n"; exit 1; }
+  $_CROSS_COMPRESS_TAR "$_CROSS_PACKAGE_DIR/$packagename$_CROSS_COMPRESS_EXT" ./* > "$logdir/package$packagesuffix.log" 2>&1 \
+    || { printf "Failure packaging $project$packagesuffix. Check $logdir/install$packagesuffix.log for details.\n"; exit 1; }
 )
